FIX: Version 0.3

+++ install.sh
@@ -46,41 +46,44 @@ then
     fi
 fi
 
-echo Creating initial password file
-touch $HOME/.pasfile.csv
-if [ $? -gt 0 ]
+if [ ! -f $HOME/.pasfile.csv.gpg ] && [ ! -f $HOME/.pasfile.csv ]
 then
-    echo COULD NOT CREATE PASSWORD FILE
-    echo QUITTING
-    sudo rm /opt/steelbox.sh
-    sudo rm /opt/steelbox.py
-    sudo rm /usr/bin/steelbox
-    exit 1
-fi
-echo Setting up password file
-echo service,user,pswd > $HOME/.pasfile.csv &> /dev/null
+    echo Creating initial password file
+    touch $HOME/.pasfile.csv
+    if [ $? -gt 0 ]
+    then
+        echo COULD NOT CREATE PASSWORD FILE
+        echo QUITTING
+        sudo rm /opt/steelbox.sh
+        sudo rm /opt/steelbox.py
+        sudo rm /usr/bin/steelbox
+        exit 1
+    fi
+    echo Setting up password file
+    echo service,user,pswd > $HOME/.pasfile.csv &> /dev/null
 
-echo Loading GPG to encrypt the file for the first time.
-echo ===========================================================
-echo YOU WILL BE ASKED TO GIVE A PASSWORD TO THE PASSWORD FILE
-echo \(And yes, I\'m not immune to the irony\)
-echo YOU WILL NEED THIS PASSWORD TO OPEN YOUR FILE
-echo gpg-agent \(OR WHATEVER AGENT YOU USE\) WILL
-echo HANDLE YOUR PASSWORDS UNTIL YOU REBOOT
-echo ===========================================================
-echo PRESS ENTER TO CONTINUE
-read
+    echo Loading GPG to encrypt the file for the first time.
+    echo ===========================================================
+    echo YOU WILL BE ASKED TO GIVE A PASSWORD TO THE PASSWORD FILE
+    echo \(And yes, I\'m not immune to the irony\)
+    echo YOU WILL NEED THIS PASSWORD TO OPEN YOUR FILE
+    echo gpg-agent \(OR WHATEVER AGENT YOU USE\) WILL
+    echo HANDLE YOUR PASSWORDS UNTIL YOU REBOOT
+    echo ===========================================================
+    echo PRESS ENTER TO CONTINUE
+    read
 
-gpg -c --cipher-algo AES256 $HOME/.pasfile.csv
-if [ $? -gt 0 ]
-then
-    echo ERROR ENCRYPTING THE FILE
-    echo YOU MUST TYPE A PASSWORD FOR THE INITIAL PASSWORD FILE
-    echo DELETING ALL INSTALATION FILES
-    sudo rm /opt/steelbox.sh
-    sudo rm /opt/steelbox.py
-    sudo rm /usr/bin/steelbox
+    gpg -c --cipher-algo AES256 $HOME/.pasfile.csv
+    if [ $? -gt 0 ]
+    then
+        echo ERROR ENCRYPTING THE FILE
+        echo YOU MUST TYPE A PASSWORD FOR THE INITIAL PASSWORD FILE
+        echo DELETING ALL INSTALATION FILES
+        sudo rm /opt/steelbox.sh
+        sudo rm /opt/steelbox.py
+        sudo rm /usr/bin/steelbox
+        rm $HOME/.pasfile.csv
+        exit 1
+    fi
     rm $HOME/.pasfile.csv
-    exit 1
 fi
-rm $HOME/.pasfile.csv
